feature/3-2 메인 페이지 컴포넌트 분류 및 마크업, 스타일링 적용(SCSS)
