+++ src/components/Layout.tsx
@@ -1,4 +1,5 @@
 import { ReactNode } from "react"
+import MenuList from "./Menu"
 
 interface LayoutProps {
   children: ReactNode
@@ -8,6 +9,7 @@ export const Layout = ({children}: LayoutProps) => {
   return (
     <div className="layout">
       {children}
+      <MenuList />
     </div>
   )
 }
